Handle 404 errors from test API call
When checking if our OAUTH token is valid, we try to retrieve the
supposedly authenticated user's profile. If this fails with a 401
(unauthorized), we know that the token we have is invalid and we need a
new one.

In some configurations, the Canvas API won't return a 401 if our
credentials are invalid. Instead, it returns a 404. This is probably for
security reasons, it prevents unauthorized users from getting side
channel data about your users.

This change watches for all status codes which are not 200 from the
Canvas API. This means that we will redirect to authentication if the
Canvas API returns a 500,  which probably means that the server
is malfunctioning. I believe this is an acceptable risk since it is
unlikely that people will reach the LTI if the server is unable to
return their profile.